+++ test/test_prototype_transforms_functional.py
@@ -1,4 +1,3 @@
-import functools
 import itertools
 import math
 import os
@@ -6,164 +5,16 @@ import os
 import numpy as np
 import PIL.Image
 import pytest
+import torch
 import torch.testing
 import torchvision.prototype.transforms.functional as F
 from common_utils import cpu_and_gpu
+from prototype_common_utils import make_bounding_boxes, make_image, make_images, make_segmentation_masks
 from torch import jit
-from torch.nn.functional import one_hot
 from torchvision.prototype import features
 from torchvision.prototype.transforms.functional._geometry import _center_crop_compute_padding
 from torchvision.prototype.transforms.functional._meta import convert_bounding_box_format
 from torchvision.transforms.functional import _get_perspective_coeffs
-from torchvision.transforms.functional_tensor import _max_value as get_max_value
-
-make_tensor = functools.partial(torch.testing.make_tensor, device="cpu")
-
-
-def make_image(size=None, *, color_space, extra_dims=(), dtype=torch.float32, constant_alpha=True):
-    size = size or torch.randint(16, 33, (2,)).tolist()
-
-    try:
-        num_channels = {
-            features.ColorSpace.GRAY: 1,
-            features.ColorSpace.GRAY_ALPHA: 2,
-            features.ColorSpace.RGB: 3,
-            features.ColorSpace.RGB_ALPHA: 4,
-        }[color_space]
-    except KeyError as error:
-        raise pytest.UsageError() from error
-
-    shape = (*extra_dims, num_channels, *size)
-    max_value = get_max_value(dtype)
-    data = make_tensor(shape, low=0, high=max_value, dtype=dtype)
-    if color_space in {features.ColorSpace.GRAY_ALPHA, features.ColorSpace.RGB_ALPHA} and constant_alpha:
-        data[..., -1, :, :] = max_value
-    return features.Image(data, color_space=color_space)
-
-
-make_grayscale_image = functools.partial(make_image, color_space=features.ColorSpace.GRAY)
-make_rgb_image = functools.partial(make_image, color_space=features.ColorSpace.RGB)
-
-
-def make_images(
-    sizes=((16, 16), (7, 33), (31, 9)),
-    color_spaces=(
-        features.ColorSpace.GRAY,
-        features.ColorSpace.GRAY_ALPHA,
-        features.ColorSpace.RGB,
-        features.ColorSpace.RGB_ALPHA,
-    ),
-    dtypes=(torch.float32, torch.uint8),
-    extra_dims=((), (0,), (4,), (2, 3), (5, 0), (0, 5)),
-):
-    for size, color_space, dtype in itertools.product(sizes, color_spaces, dtypes):
-        yield make_image(size, color_space=color_space, dtype=dtype)
-
-    for color_space, dtype, extra_dims_ in itertools.product(color_spaces, dtypes, extra_dims):
-        yield make_image(size=sizes[0], color_space=color_space, extra_dims=extra_dims_, dtype=dtype)
-
-
-def randint_with_tensor_bounds(arg1, arg2=None, **kwargs):
-    low, high = torch.broadcast_tensors(
-        *[torch.as_tensor(arg) for arg in ((0, arg1) if arg2 is None else (arg1, arg2))]
-    )
-    return torch.stack(
-        [
-            torch.randint(low_scalar, high_scalar, (), **kwargs)
-            for low_scalar, high_scalar in zip(low.flatten().tolist(), high.flatten().tolist())
-        ]
-    ).reshape(low.shape)
-
-
-def make_bounding_box(*, format, image_size=(32, 32), extra_dims=(), dtype=torch.int64):
-    if isinstance(format, str):
-        format = features.BoundingBoxFormat[format]
-
-    if any(dim == 0 for dim in extra_dims):
-        return features.BoundingBox(torch.empty(*extra_dims, 4), format=format, image_size=image_size)
-
-    height, width = image_size
-
-    if format == features.BoundingBoxFormat.XYXY:
-        x1 = torch.randint(0, width // 2, extra_dims)
-        y1 = torch.randint(0, height // 2, extra_dims)
-        x2 = randint_with_tensor_bounds(x1 + 1, width - x1) + x1
-        y2 = randint_with_tensor_bounds(y1 + 1, height - y1) + y1
-        parts = (x1, y1, x2, y2)
-    elif format == features.BoundingBoxFormat.XYWH:
-        x = torch.randint(0, width // 2, extra_dims)
-        y = torch.randint(0, height // 2, extra_dims)
-        w = randint_with_tensor_bounds(1, width - x)
-        h = randint_with_tensor_bounds(1, height - y)
-        parts = (x, y, w, h)
-    elif format == features.BoundingBoxFormat.CXCYWH:
-        cx = torch.randint(1, width - 1, ())
-        cy = torch.randint(1, height - 1, ())
-        w = randint_with_tensor_bounds(1, torch.minimum(cx, width - cx) + 1)
-        h = randint_with_tensor_bounds(1, torch.minimum(cy, height - cy) + 1)
-        parts = (cx, cy, w, h)
-    else:
-        raise pytest.UsageError()
-
-    return features.BoundingBox(torch.stack(parts, dim=-1).to(dtype), format=format, image_size=image_size)
-
-
-make_xyxy_bounding_box = functools.partial(make_bounding_box, format=features.BoundingBoxFormat.XYXY)
-
-
-def make_bounding_boxes(
-    formats=(features.BoundingBoxFormat.XYXY, features.BoundingBoxFormat.XYWH, features.BoundingBoxFormat.CXCYWH),
-    image_sizes=((32, 32),),
-    dtypes=(torch.int64, torch.float32),
-    extra_dims=((0,), (), (4,), (2, 3), (5, 0), (0, 5)),
-):
-    for format, image_size, dtype in itertools.product(formats, image_sizes, dtypes):
-        yield make_bounding_box(format=format, image_size=image_size, dtype=dtype)
-
-    for format, extra_dims_ in itertools.product(formats, extra_dims):
-        yield make_bounding_box(format=format, extra_dims=extra_dims_)
-
-
-def make_label(size=(), *, categories=("category0", "category1")):
-    return features.Label(torch.randint(0, len(categories) if categories else 10, size), categories=categories)
-
-
-def make_one_hot_label(*args, **kwargs):
-    label = make_label(*args, **kwargs)
-    return features.OneHotLabel(one_hot(label, num_classes=len(label.categories)), categories=label.categories)
-
-
-def make_one_hot_labels(
-    *,
-    num_categories=(1, 2, 10),
-    extra_dims=((), (0,), (4,), (2, 3), (5, 0), (0, 5)),
-):
-    for num_categories_ in num_categories:
-        yield make_one_hot_label(categories=[f"category{idx}" for idx in range(num_categories_)])
-
-    for extra_dims_ in extra_dims:
-        yield make_one_hot_label(extra_dims_)
-
-
-def make_segmentation_mask(size=None, *, num_objects=None, extra_dims=(), dtype=torch.uint8):
-    size = size if size is not None else torch.randint(16, 33, (2,)).tolist()
-    num_objects = num_objects if num_objects is not None else int(torch.randint(1, 11, ()))
-    shape = (*extra_dims, num_objects, *size)
-    data = make_tensor(shape, low=0, high=2, dtype=dtype)
-    return features.SegmentationMask(data)
-
-
-def make_segmentation_masks(
-    sizes=((16, 16), (7, 33), (31, 9)),
-    dtypes=(torch.uint8,),
-    extra_dims=((), (0,), (4,), (2, 3), (5, 0), (0, 5)),
-    num_objects=(1, 0, 10),
-):
-    for size, dtype, extra_dims_ in itertools.product(sizes, dtypes, extra_dims):
-        yield make_segmentation_mask(size=size, dtype=dtype, extra_dims=extra_dims_)
-
-    for dtype, extra_dims_, num_objects_ in itertools.product(dtypes, extra_dims, num_objects):
-        yield make_segmentation_mask(num_objects=num_objects_, dtype=dtype, extra_dims=extra_dims_)
 
 
 class SampleInput:
@@ -815,12 +666,7 @@ def test_correctness_affine_bounding_box(angle, translate, scale, shear, center)
 
     image_size = (32, 38)
 
-    for bboxes in make_bounding_boxes(
-        image_sizes=[
-            image_size,
-        ],
-        extra_dims=((4,),),
-    ):
+    for bboxes in make_bounding_boxes(image_size=image_size, extra_dims=((4,),)):
         bboxes_format = bboxes.format
         bboxes_image_size = bboxes.image_size
 
@@ -1038,12 +884,7 @@ def test_correctness_rotate_bounding_box(angle, expand, center):
 
     image_size = (32, 38)
 
-    for bboxes in make_bounding_boxes(
-        image_sizes=[
-            image_size,
-        ],
-        extra_dims=((4,),),
-    ):
+    for bboxes in make_bounding_boxes(image_size=image_size, extra_dims=((4,),)):
         bboxes_format = bboxes.format
         bboxes_image_size = bboxes.image_size
 
@@ -1587,12 +1428,7 @@ def test_correctness_perspective_bounding_box(device, startpoints, endpoints):
     pcoeffs = _get_perspective_coeffs(startpoints, endpoints)
     inv_pcoeffs = _get_perspective_coeffs(endpoints, startpoints)
 
-    for bboxes in make_bounding_boxes(
-        image_sizes=[
-            image_size,
-        ],
-        extra_dims=((4,),),
-    ):
+    for bboxes in make_bounding_boxes(image_size=image_size, extra_dims=((4,),)):
         bboxes = bboxes.to(device)
         bboxes_format = bboxes.format
         bboxes_image_size = bboxes.image_size
@@ -1714,10 +1550,7 @@ def test_correctness_center_crop_bounding_box(device, output_size):
         )
         return convert_bounding_box_format(out_bbox, features.BoundingBoxFormat.XYWH, format_, copy=False)
 
-    for bboxes in make_bounding_boxes(
-        image_sizes=[(32, 32), (24, 33), (32, 25)],
-        extra_dims=((4,),),
-    ):
+    for bboxes in make_bounding_boxes(extra_dims=((4,),)):
         bboxes = bboxes.to(device)
         bboxes_format = bboxes.format
         bboxes_image_size = bboxes.image_size
